Фиксы по реквестам SNMetamorph-а
