+++ server/entities/gargantua_stomp.h
@@ -29,7 +29,7 @@
 #include	"weapons.h"
 #include	"effects.h"
 #include	"decals.h"
-#include	"explode.h"
+#include	"env_explosion.h"
 #include	"func_break.h"
 
 #define GARG_STOMP_SPRITE_NAME	"sprites/gargeye1.spr"
