+++ server/entities/path_corner.h
