+++ server/monsters/nodes.cpp
@@ -23,7 +23,7 @@
 #include	"monsters.h"
 #include	"nodes.h"
 #include	"animation.h"
-#include	"doors.h"
+#include	"func_door.h"
 
 #define	HULL_STEP_SIZE 16// how far the test hull moves on each step
 #define	NODE_HEIGHT	8	// how high to lift nodes off the ground after we drop them all (make stair/ramp mapping easier)
